Move wait-for-it.sh to $PATH and make executable.
Given that this script is used internally as part of the docker image, it makes sense to ensure it is added to the path directly within the Dockerfile itself, rather than from the docker-compose.yml file. It also needs to be made executable, otherwise the entrypoint command within the docker-compose.yml file will not be able to run it.