Remove wait-for-it from docker-compose.yml
Given that the script is used internally, it should be added to the Dockerfile, not the docker-compose.yml. Trying to include it from the docker-compose.yml file means that you might run into issues where it is not executable, and therefore the entrypoint command cannot run it (as it won't exist as an executable).